fix: resolve all 114 mypy strict errors; credit Peixoto lab in README
- pyproject.toml: add [[tool.mypy.overrides]] ignore_missing_imports for
  networkx, PyYAML, langchain-*, igraph, leidenalg, sklearn, graph_tool,
  docling, anthropic, datasets, scipy, rank_bm25
- Remove stale # type: ignore comments in hub_exclusion.py and graph/io.py
- Add missing type annotations across ingest, eval, viz, semantic_extract
- Fix no-any-return in source_metadata, publish, embed, dedup, ragas_runner
- Fix type-arg (bare dict/Counter) in protocol, sbm, persist, eval modules
- Fix incompatible assignment types in obsidian.py and cli.py
- README: expand Peixoto/graph-tool attribution with lab link and affiliation

+++ src/nuthatch/dedup/semantic.py
@@ -90,7 +90,7 @@ def _cosine(a: Sequence[float], b: Sequence[float]) -> float:
     db = sum(y * y for y in b) ** 0.5
     if da == 0.0 or db == 0.0:
         return 0.0
-    return num / (da * db)
+    return float(num) / float(da * db)
 
 
 def classify(
